+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v38.x / v39 (In progress)
 - Ancient Cave: All previously-disabled one-way doors are now enabled, and properly two-way. This will cause major routing changes.
+- Ancient Cave: Loops may now appear within a floor.
 - PC sprites: Added Dog (Laika).
 
 ## v38 (2019-05-10)